+++ src/@tests/functional/mongo/models/seeders.test.ts
@@ -3,60 +3,72 @@
 import { assert, assertEquals } from '@std/assert'
 import { DropCollection, getDB, ignore, sanitize } from '../../../_setup/mongo/connector.ts'
 import { defineModelHOC } from 'modules/database/hocs/models.ts'
+import { seedByIdIfMissing, seedManyByIdIfMissing } from 'mongo/utils/seeders.ts'
 import { Schema } from 'mongoose'
 
+const bulk = [
+  {
+    id: '68fb00b33405a3a540d9b971',
+    name: 'test-seeder-1',
+    description: 'description test seeder 1',
+  },
+  {
+    id: '68fb00b33405a3a540d9b972',
+    name: 'test-seeder-2',
+    description: 'description test seeder not saved',
+  },
+  {
+    id: '68fb00b33405a3a540d9b973',
+    name: 'test-seeder-3',
+    description: 'description test seeder 3',
+  },
+]
 export const extensionsSeeders = {
   seeders: [
     async (Model: any) => {
-      const data = await Model.findById('68fb00b33405a3a540d9b971')
+      const data = await Model.findById(bulk[0].id)
       if (data) return
-      const acc = new Model({
-        id: '68fb00b33405a3a540d9b971',
-        name: 'test-seeder-1',
-        description: 'description test seeder 1',
-      })
+      const acc = new Model(bulk[0])
 
       return acc.save()
     },
     async (Model: any) => {
-      const data = await Model.findById('68fb00b33405a3a540d9b971')
+      const data = await Model.findById(bulk[0].id)
       if (data) return
-      const acc = new Model({
-        id: '68fb00b33405a3a540d9b972',
-        name: 'test-seeder-2',
-        description: 'description test seeder not saved',
-      })
+      const acc = new Model(bulk[1])
 
       return acc.save()
     },
     async (Model: any) => {
-      const data = await Model.findById('68fb00b33405a3a540d9b973')
+      const data = await Model.findById(bulk[2].id)
       if (data) return
-      const acc = new Model({
-        id: '68fb00b33405a3a540d9b973',
-        name: 'test-seeder-3',
-        description: 'description test seeder 3',
-      })
+      const acc = new Model(bulk[2])
 
       return acc.save()
     },
   ],
 }
 
+const closeConnection = async (Model: any, db: any) => {
+  // Drop collection
+  await Model.deleteMany({ _id: { $in: [bulk[0].id, bulk[1].id, bulk[2].id] } })
+  await DropCollection(Model, db)
+  await db['stopConnection']()
+}
+
 const seedersValidation = async (Model: any, db: any) => {
   const [seeder1, seeder2, seeder3] = await Promise.all([
-    Model.findById('68fb00b33405a3a540d9b971'),
-    Model.findById('68fb00b33405a3a540d9b972'),
-    Model.findById('68fb00b33405a3a540d9b973'),
+    Model.findById(bulk[0].id),
+    Model.findById(bulk[1].id),
+    Model.findById(bulk[2].id),
   ])
 
   assert(!seeder2)
-  assertEquals(seeder1?.name, 'test-seeder-1')
-  assertEquals(seeder3?.name, 'test-seeder-3')
+  assertEquals(seeder1?.name, bulk[0].name)
+  assertEquals(seeder3?.name, bulk[2].name)
 
   // Drop collection
-  await DropCollection(Model, db)
-  await db['stopConnection']()
+  await closeConnection(Model, db)
 }
 
 Deno.test({
@@ -97,6 +109,41 @@ Deno.test({
     })
 
     await seedersValidation(Model, db)
+  },
+  ignore,
+})
+
+Deno.test({
+  ...sanitize,
+  name: 'Mongo connector should run customized seeders',
+  fn: async () => {
+    defineModelHOC({
+      name: 'test-customized-seeders',
+      definition: {
+        name: String,
+        description: String,
+      },
+      extensions: {
+        seeders: [
+          seedByIdIfMissing(bulk[0]),
+          seedManyByIdIfMissing([bulk[1], bulk[2]]),
+          seedByIdIfMissing(bulk[0]),
+        ],
+      },
+    })
+
+    const db = await getDB()
+    const Model = db.getModel('test-customized-seeders')
+
+    const all = await Promise.all([
+      Model.findById(bulk[0].id),
+      Model.findById(bulk[1].id),
+      Model.findById(bulk[2].id),
+    ])
+
+    assert(!all.some((data) => !data))
+
+    await closeConnection(Model, db)
   },
   ignore,
 })