+++ docs/CHANGELOG.md
@@ -7,6 +7,14 @@ adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.1.5] - 2025-11-04
+
+### Added
+
+- Mongo testing support on CI
+- Mongo base seeders
+- Some mongo model statics
+
 ## [0.1.4] - 2025-11-04
 
 ### Added